pyproject.toml
[project]
name = "qa-generator"
version = "0.1.0"
description = "QA Dataset Generator"
requires-python = ">=3.13"
dependencies = [
    "export>=0.2.1",
    "google-genai>=1.24.0",
    "json-repair>=0.47.6",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "pydantic-settings>=2.10.1",
    "pip-audit>=2.9.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
